+++ src/PowerUtils.Validations/Contracts/PropertyRule.cs
@@ -1,10 +1,28 @@
-﻿using System.Collections.Generic;
+﻿#region License
+// Copyright (c) .NET Foundation and contributors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// The latest version of this file can be found at https://github.com/FluentValidation/FluentValidation
+#endregion
+
+
+using System.Collections.Generic;
 
 namespace PowerUtils.Validations.Contracts
 {
     public class PropertyRule<TSource, TProperty> : IPropertyRule<TSource, TProperty>
     {
-        #region PROPERTIES
         public bool PropertyNull { get; init; }
         public string PropertyName { get; init; }
         public TProperty PropertyValue { get; init; }
@@ -17,10 +35,8 @@ namespace PowerUtils.Validations.Contracts
 
 
         private IValidationsContract<TSource> _context { get; }
-        #endregion
 
 
-        #region CONSTRUCTORS
         public PropertyRule(
             IValidationsContract<TSource> context,
             string propertyName
@@ -47,7 +63,6 @@ namespace PowerUtils.Validations.Contracts
                 PropertyNull = false;
             }
         }
-        #endregion
 
 
         #region METHODS - NOTIFICATIONS